+++ CHANGES.rst
@@ -26,7 +26,7 @@ v3.3 (2021-03-04)
 * :issue:`504`: Fix a crash on Windows.
 * :issue:`490`: When ``--rename`` is used with ``--revcomp``, disable adding the
   ``rc`` suffix to reads that were reverse-complemented.
-* Also, there is now a ``{rc}` template variable for the ``--rename`` option, which
+* Also, there is now a ``{rc}`` template variable for the ``--rename`` option, which
   is replaced with "rc" if the read was reverse-complemented (and the empty string if not).
 * :issue:`512`: Fix issue :issue:`128` once more (the “Reads written” figure in the report
   incorrectly included both trimmed and untrimmed reads if ``--untrimmed-output`` was used).